move Instabug token to gradle.propertiess
