+++ app/src/playstoreDebug/java/tech/schober/vinylcast/VinylCastApplicationBase.java
@@ -2,9 +2,16 @@ package tech.schober.vinylcast;
 
 import android.app.Application;
 import android.os.StrictMode;
+import android.util.Log;
 
+import com.google.firebase.crashlytics.FirebaseCrashlytics;
 import com.instabug.library.Instabug;
 import com.instabug.library.invocation.InstabugInvocationEvent;
+
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import timber.log.Timber;
 
 public class VinylCastApplicationBase extends Application {
     private static final String TAG = "VinylCastApplicationBase";
@@ -30,8 +37,27 @@ public class VinylCastApplicationBase extends Application {
                     .build());
         }
 
-        new Instabug.Builder(this, "40d0b4f618b8d8eacd26fceff152a822")
+        new Instabug.Builder(this, BuildConfig.INSTABUG_TOKEN)
                 .setInvocationEvents(InstabugInvocationEvent.NONE)
                 .build();
+
+        if (BuildConfig.DEBUG) {
+            Timber.plant(new Timber.DebugTree());
+        }
+        Timber.plant(new CrashReportingTree());
+    }
+
+    private class CrashReportingTree extends Timber.Tree {
+        @Override
+        protected void log(int priority, @Nullable String tag, @NotNull String message, @Nullable Throwable throwable) {
+            if (priority >= Log.DEBUG) {
+                FirebaseCrashlytics.getInstance().log(message);
+                if (throwable != null) {
+                    FirebaseCrashlytics.getInstance().recordException(throwable);
+                }
+            } else {
+                return;
+            }
+        }
     }
 }
