+++ gradle.properties
@@ -18,3 +18,7 @@
 # org.gradle.parallel=true
 android.enableJetifier=true
 android.useAndroidX=true
+
+INSTABUG_TOKEN_BETA=""
+INSTABUG_TOKEN_LIVE=""
+
